feat/ User Authorisation management

+++ users-service/src/graphql/resolvers.js
@@ -1,6 +1,8 @@
 const { GraphQLObjectType, GraphQLString, GraphQLList } = require("graphql");
 const { getAllUsers, getUserByEmail, updateUser, deleteUser } = require("../repositories/users-repository");
 const UserType = require("./types");
+const { register, login } = require("../repositories/auth-repository");
+const { GraphQLNonNull } = require("graphql/type");
 
 const RootQuery = new GraphQLObjectType({
     name: 'Query',
@@ -12,7 +14,13 @@ const RootQuery = new GraphQLObjectType({
         getUserByEmail: {
             type: UserType,
             args: { email: { type: GraphQLString } },
-            resolve: async (_, { email }) => await getUserByEmail(email),
+            resolve: async (_, { email }, { user }) => {
+                // Create custom object to compare email as users id
+                const userToCompare = user ? { id: user.email, role: user.role } : undefined
+                checkUserAuthorizations(email, userToCompare)
+
+                return await getUserByEmail(email)
+            },
         }
     }
 })
@@ -23,29 +31,63 @@ const Mutations = new GraphQLObjectType({
         updateUser: {
             type: UserType,
             args: {
-                id: {type: GraphQLString},
-                email: {type: GraphQLString},
-                pseudo: {type: GraphQLString},
-                password: {type: GraphQLString}
+                id: { type: new GraphQLNonNull(GraphQLString) },
+                email: { type: GraphQLString },
+                pseudo: { type: GraphQLString },
+                password: { type: GraphQLString }
             },
-            resolve: async (_, args) => {
-                // if (!user) throw new Error("Unauthorized");
-                // if (user.id !== args.id) throw new Error("You can only update your own profile");
-
-                return await updateUser(args.id, args);
+            resolve: async (_, args, { user }) => {
+                checkUserAuthorizations(args.id, user)
+                return await updateUser(args.id, args)
             }
         },
         deleteUser: {
             type: UserType,
             args: { id: { type: GraphQLString } },
-            resolve: async (_, { id }) => {
-                // if (!user) throw new Error("Unauthorized");
-                // if (user.id !== id) throw new Error("You can only delete your own account");
-
+            resolve: async (_, { id }, { user }) => {
+                checkUserAuthorizations(id, user)
                 return await deleteUser(id);
+            }
+        },
+        register: {
+            type: UserType,
+            args: {
+                email: { type: new GraphQLNonNull(GraphQLString) },
+                pseudo: { type: new GraphQLNonNull(GraphQLString) },
+                password: { type: new GraphQLNonNull(GraphQLString) },
+                role: { type: new GraphQLNonNull(GraphQLString) },
+            },
+            resolve: async (_, { email, pseudo, password, role }) => {
+                return await register({ email, pseudo, password, role })
+            }
+        },
+        login: {
+            type: new GraphQLObjectType({
+                name: 'AuthPayload',
+                fields: {
+                    token: { type: GraphQLString },
+                    user: { type: UserType },
+                }
+            }),
+            args: {
+                email: { type: new GraphQLNonNull(GraphQLString) },
+                password: { type: new GraphQLNonNull(GraphQLString) },
+            },
+            resolve: async (_, { email, password }) => {
+                return await login({ email, password })
             }
         }
     }
 })
+
+function checkUserAuthorizations(askedId, user) {
+    if (!user) {
+        throw new Error("Unauthorized")
+    }
+
+    if (user.id !== askedId && user.role !== "ROLE_ADMIN") {
+        throw new Error("You can only manage your own account")
+    }
+}
 
 module.exports = { RootQuery, Mutations }
